requirements.txt
streamlit
google-generativeai
requests
pandas
python-dotenv
scikit-learn==1.3.2
joblib
beautifulsoup4
xgboost
lightgbm

numpy
